refactor(canvas): establish chat view module
