+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatView/index.tsx
@@ -13,16 +13,16 @@ import type {
   CanvasNode,
   ChatImageAttachment,
   RelayProgress,
-} from '../../types';
+} from '../../../types';
 import { ChatEmptyState, type ChatEmptyStateVariant } from './ChatEmptyState';
-import { ChatInput } from './ChatInput';
-import { ChatMentionPopup } from './ChatMentionPopup';
-import { ChatMessages } from './ChatMessages';
+import { ChatInput } from '../ChatInput';
+import { ChatMentionPopup } from '../ChatMentionPopup';
+import { ChatMessages } from '../ChatMessages';
 import { RelayBar } from './RelayBar';
-import type { QueuedInput } from '../../agent-chat/runtime/useChatRunQueue';
-import type { MentionItem, PendingClarification, SelectedContextChip, ToolCallStatus } from './types';
-import { restoreComposerFocusAfterRender } from './utils/focusRecovery';
-import { McpAppsProvider } from '../mcp-apps/McpAppsProvider';
+import type { QueuedInput } from '../../../agent-chat/runtime/useChatRunQueue';
+import type { MentionItem, PendingClarification, SelectedContextChip, ToolCallStatus } from '../types';
+import { restoreComposerFocusAfterRender } from '../utils/focusRecovery';
+import { McpAppsProvider } from '../../mcp-apps/McpAppsProvider';
 
 interface ChatViewProps {
   className?: string;
