+++ apps/canvas-workspace/src/renderer/src/components/chat/__tests__/ChatEmptyState.test.tsx
@@ -3,7 +3,7 @@ import { act } from 'react';
 import { createRoot, type Root } from 'react-dom/client';
 import { afterEach, describe, expect, it, vi } from 'vitest';
 import { I18nProvider } from '../../../i18n';
-import { ChatEmptyState } from '../ChatEmptyState';
+import { ChatEmptyState } from '../ChatView/ChatEmptyState';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
